Sorting via fixed sprite and a given offset implemented
Instead of giving every gameobject a sorting point (GameObject) you can now add a sprite and a fixed yOffset to a list. Every registered moving SpriteRender with the given Sprite will use the associated yOffset of the list to sort the sprite order. This will be applied to static objects too. Important: If a GameObject does have a Sorting Point attached as a child object, this sorting point will be selected as a sorting reference